pyproject.toml
[project]
name = "aibingoslop"
version = "0.1.0"
description = "CLI to generate Bingo Chart on AI Slop Startup"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pillow>=11.3.0",
  "pyright>=1.1.405",
  "reportlab>=4.4.4",
  "rich>=14.1.0",
  "typer>=0.19.2",
]

[tool.ruff.format]

[project.scripts]
aibingoslop = "aibingoslop.cli:app"
